+++ Source/AccelByteUe4Sdk/AccelByteUe4Sdk.Build.cs
@@ -60,7 +60,11 @@ public class AccelByteUe4Sdk : ModuleRules
 		{
 			PrivateDependencyModuleNames.AddRange(new[]
 			{
-				"Settings"
+				"Settings",
+				"ToolMenus",
+				"UnrealEd",
+				"Slate",
+				"SlateCore"
 			});
 		}